Fix sg28xx panel refresh rate
This change fixes the sun4-tcon driver so that it not includes the sync
phase into the back porch value. This confuses the panel of the sg28xx so
that it does not properly render content at 60Hz.

Secondly this change relaxes the timing a bit as it was causing slight
horizontal artifacts when running at 60 Hz. Also the Porch control register
of st7701 is set to properly work in DE mode (Data enable RGB mode).

Signed-off-by: Hendrik Borghorst <hendrikborghorst@gmail.com>